# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Formula 1 API",
    "version": "1.0.0",
    "description": "API for managing Formula 1 team and driver data",
    "contact": {
      "name": "Petra Habjanec",
      "email": "petra.habjanec@fer.hr"
    },
    "license": {
      "name": "CC0",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/legalcode"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/v1/data": {
      "get": {
        "summary": "Get all team data with drivers",
        "description": "Returns all the team data and its drivers",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "teamName": {
                            "type": "string"
                          },
                          "teamBase": {
                            "type": "string"
                          },
                          "teamChief": {
                            "type": "string"
                          },
                          "chassis": {
                            "type": "string"
                          },
                          "powerUnit": {
                            "type": "string"
                          },
                          "firstEntry": {
                            "type": "number"
                          },
                          "teamChampionships": {
                            "type": "number"
                          },
                          "driverChampionships": {
                            "type": "number"
                          },
                          "highestRaceFinish": {
                            "type": "number"
                          },
                          "polePositions": {
                            "type": "number"
                          },
                          "fastestLaps": {
                            "type": "number"
                          },
                          "points": {
                            "type": "number"
                          },
                          "drivers": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "driverName": {
                                  "type": "string"
                                },
                                "driverLast": {
                                  "type": "string"
                                },
                                "driverNum": {
                                  "type": "number"
                                },
                                "nationality": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Empty database response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/{id}": {
      "get": {
        "summary": "Get team data with drivers by ID",
        "description": "Return a single team with its drivers",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "response": {
                      "type": "object",
                      "properties": {
                        "teamName": {
                          "type": "string"
                        },
                        "teamBase": {
                          "type": "string"
                        },
                        "teamChief": {
                          "type": "string"
                        },
                        "drivers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "teamName": {
                                "type": "string"
                              },
                              "teamBase": {
                                "type": "string"
                              },
                              "teamChief": {
                                "type": "string"
                              },
                              "chassis": {
                                "type": "string"
                              },
                              "powerUnit": {
                                "type": "string"
                              },
                              "firstEntry": {
                                "type": "number"
                              },
                              "teamChampionships": {
                                "type": "number"
                              },
                              "driverChampionships": {
                                "type": "number"
                              },
                              "highestRaceFinish": {
                                "type": "number"
                              },
                              "polePositions": {
                                "type": "number"
                              },
                              "fastestLaps": {
                                "type": "number"
                              },
                              "points": {
                                "type": "number"
                              },
                              "drivers": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "driverName": {
                                      "type": "string"
                                    },
                                    "driverLast": {
                                      "type": "string"
                                    },
                                    "driverNum": {
                                      "type": "number"
                                    },
                                    "nationality": {
                                      "type": "string"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error response - Wrong ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team/{id}": {
      "get": {
        "summary": "Get team information by ID",
        "description": "Return a single team info",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "teamName": {
                            "type": "string"
                          },
                          "teamBase": {
                            "type": "string"
                          },
                          "teamChief": {
                            "type": "string"
                          },
                          "chassis": {
                            "type": "string"
                          },
                          "powerUnit": {
                            "type": "string"
                          },
                          "firstEntry": {
                            "type": "number"
                          },
                          "teamChampionships": {
                            "type": "number"
                          },
                          "driverChampionships": {
                            "type": "number"
                          },
                          "highestRaceFinish": {
                            "type": "number"
                          },
                          "polePositions": {
                            "type": "number"
                          },
                          "fastestLaps": {
                            "type": "number"
                          },
                          "points": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error response - Wrong ID or data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/driver/{id}": {
      "get": {
        "summary": "Get driver information by ID",
        "description": "Return a single driver info",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Driver ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "driverNum": {
                            "type": "number"
                          },
                          "driverName": {
                            "type": "string"
                          },
                          "driverLast": {
                            "type": "string"
                          },
                          "nationality": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error response - Wrong ID or data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drivers/{id}": {
      "get": {
        "summary": "Get drivers for a team by ID",
        "description": "Return drivers info by their team id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "driverNum": {
                            "type": "number"
                          },
                          "driverName": {
                            "type": "string"
                          },
                          "driverLast": {
                            "type": "string"
                          },
                          "nationality": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error response - Wrong ID or data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/delete/{id}": {
      "delete": {
        "summary": "Delete an existing team",
        "description": "Deletes team by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Team name (ID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error response - Team does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}